Complete generic Steamworks API coverage

+++ steam_api/Steamworks/Interfaces/SteamUGC/SteamUGC015.cs
@@ -26,7 +26,7 @@ namespace SKYNET.Steamworks.Interfaces
             return SteamEmulator.SteamUGC.CreateQueryAllUGCRequest(eQueryType, eMatchingeMatchingUGCTypeFileType, nCreatorAppID, nConsumerAppID, pchCursor);
         }
 
-        public UGCQueryHandle_t CreateQueryUGCDetailsRequest(IntPtr _, ulong pvecPublishedFileID, uint unNumPublishedFileIDs)
+        public UGCQueryHandle_t CreateQueryUGCDetailsRequest(IntPtr _, IntPtr pvecPublishedFileID, uint unNumPublishedFileIDs)
         {
             return SteamEmulator.SteamUGC.CreateQueryUGCDetailsRequest(pvecPublishedFileID, unNumPublishedFileIDs);
         }
@@ -341,7 +341,7 @@ namespace SKYNET.Steamworks.Interfaces
         return SteamEmulator.SteamUGC.GetNumSubscribedItems();
     }
 
-    public uint GetSubscribedItems(IntPtr _, ulong pvecPublishedFileID, uint cMaxEntries)  // all subscribed item PublishFileIDs
+    public uint GetSubscribedItems(IntPtr _, IntPtr pvecPublishedFileID, uint cMaxEntries)  // all subscribed item PublishFileIDs
     {
         return SteamEmulator.SteamUGC.GetSubscribedItems(pvecPublishedFileID, cMaxEntries);
     }
